cleanup(phase8): move export_collection_task to KB domain (Option β follow-up)
Carve-finishing follow-up to #1670 (task #39): now that ExportTask ORM
lives in aperag/domains/knowledge_base/db/models.py, the matching Celery
task body can move to the KB domain too.

- Move `export_collection_task` + `_build_manifest` from
  `config/export_tasks.py` to `aperag/domains/knowledge_base/tasks.py`,
  pinned with `name="config.celery_tasks.export_collection_task"` to
  match the convention of the other KB-domain tasks.
- Update the only caller (`aperag.service.export_service`) to import
  from the new location.
- Drop `config.export_tasks` from `config/celery.py` `include=` list.
- Delete `config/export_tasks.py`.
- Update KB models docstring to reflect the task body's new home.

Pure code move + import path update — no behavior change. Gates green
(21/21 boundary, 676 pass / 29 skip / 1 deselect / 0 fail unit suite,
ruff clean).
